Refactor jsonrpc factories (#1)
