+++ besu/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -21,7 +21,7 @@ import static org.hyperledger.besu.controller.KeyPairUtil.loadKeyPair;
 import org.hyperledger.besu.config.GenesisConfigFile;
 import org.hyperledger.besu.crypto.SECP256K1.KeyPair;
 import org.hyperledger.besu.ethereum.ProtocolContext;
-import org.hyperledger.besu.ethereum.api.jsonrpc.method.JsonRpcMethodFactory;
+import org.hyperledger.besu.ethereum.api.jsonrpc.methods.JsonRpcMethods;
 import org.hyperledger.besu.ethereum.blockcreation.MiningCoordinator;
 import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.chain.GenesisState;
@@ -279,7 +279,7 @@ public abstract class BesuControllerBuilder<C> {
     final SubProtocolConfiguration subProtocolConfiguration =
         createSubProtocolConfiguration(ethProtocolManager);
 
-    final JsonRpcMethodFactory additionalJsonRpcMethodFactory =
+    final JsonRpcMethods additionalJsonRpcMethodFactory =
         createAdditionalJsonRpcMethodFactory(protocolContext);
 
     List<Closeable> closeables = new ArrayList<>();
@@ -307,7 +307,7 @@ public abstract class BesuControllerBuilder<C> {
 
   protected void prepForBuild() {}
 
-  protected JsonRpcMethodFactory createAdditionalJsonRpcMethodFactory(
+  protected JsonRpcMethods createAdditionalJsonRpcMethodFactory(
       final ProtocolContext<C> protocolContext) {
     return apis -> Collections.emptyMap();
   }
