+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/methods/Web3JsonRpcMethods.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.ethereum.api.jsonrpc.methods;
+
+import org.hyperledger.besu.ethereum.api.jsonrpc.RpcApi;
+import org.hyperledger.besu.ethereum.api.jsonrpc.RpcApis;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.JsonRpcMethod;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.Web3ClientVersion;
+import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.Web3Sha3;
+
+import java.util.Map;
+
+public class Web3JsonRpcMethods extends ApiGroupJsonRpcMethods {
+
+  private final String clientVersion;
+
+  public Web3JsonRpcMethods(final String clientVersion) {
+    this.clientVersion = clientVersion;
+  }
+
+  @Override
+  protected RpcApi getApiGroup() {
+    return RpcApis.WEB3;
+  }
+
+  @Override
+  protected Map<String, JsonRpcMethod> create() {
+    return mapOf(new Web3ClientVersion(clientVersion), new Web3Sha3());
+  }
+}
